requirements.txt
awesome-slugify==1.6.5
pandas==0.22.0
hdx-python-api==2.9.5
